+++ src/paserk.rs
@@ -2,10 +2,9 @@
 
 use crate::common::{decode_b64, encode_b64};
 use crate::errors::Error;
-use crate::keys::private::Version;
-use crate::keys::{
-    AsymmetricKeyPair, AsymmetricPublicKey, AsymmetricSecretKey, SymmetricKey, V2, V3, V4,
-};
+use crate::keys::{AsymmetricKeyPair, AsymmetricPublicKey, AsymmetricSecretKey, SymmetricKey};
+use crate::version::private::Version;
+use crate::{V2, V3, V4};
 use alloc::string::String;
 use alloc::vec::Vec;
 use core::convert::TryFrom;
@@ -59,7 +58,7 @@ impl TryFrom<String> for SymmetricKey<V2> {
 
     fn try_from(value: String) -> Result<Self, Self::Error> {
         Ok(Self {
-            bytes: validate_paserk_string(value.as_str(), "k2", "local", V2::LOCAL)?,
+            bytes: validate_paserk_string(value.as_str(), "k2", "local", V2::LOCAL_KEY)?,
             phantom: PhantomData,
         })
     }
@@ -77,7 +76,7 @@ impl TryFrom<String> for SymmetricKey<V4> {
 
     fn try_from(value: String) -> Result<Self, Self::Error> {
         Ok(Self {
-            bytes: validate_paserk_string(value.as_str(), "k4", "local", V4::LOCAL)?,
+            bytes: validate_paserk_string(value.as_str(), "k4", "local", V4::LOCAL_KEY)?,
             phantom: PhantomData,
         })
     }
@@ -90,9 +89,9 @@ impl FormatAsPaserk for AsymmetricKeyPair<V2> {
         // See spec: "Here, Ed25519 secret key means the clamped 32-byte seed followed by the
         // 32-byte public key, as used in the NaCl and libsodium APIs, rather than just the
         // clamped 32-byte seed."
-        let mut buf = [0u8; V2::SECRET + V2::PUBLIC];
-        buf[..V2::SECRET].copy_from_slice(self.secret.as_bytes());
-        buf[V2::SECRET..].copy_from_slice(self.public.as_bytes());
+        let mut buf = [0u8; V2::SECRET_KEY + V2::PUBLIC_KEY];
+        buf[..V2::SECRET_KEY].copy_from_slice(self.secret.as_bytes());
+        buf[V2::SECRET_KEY..].copy_from_slice(self.public.as_bytes());
         write.write_str(&encode_b64(buf).map_err(|_| core::fmt::Error)?)?;
         buf.iter_mut().zeroize();
 
@@ -104,11 +103,15 @@ impl TryFrom<String> for AsymmetricKeyPair<V2> {
     type Error = Error;
 
     fn try_from(value: String) -> Result<Self, Self::Error> {
-        let mut buf =
-            validate_paserk_string(value.as_str(), "k2", "secret", V2::SECRET + V2::PUBLIC)?;
+        let mut buf = validate_paserk_string(
+            value.as_str(),
+            "k2",
+            "secret",
+            V2::SECRET_KEY + V2::PUBLIC_KEY,
+        )?;
         let ret = Self {
-            secret: AsymmetricSecretKey::from(&buf[..V2::SECRET])?,
-            public: AsymmetricPublicKey::from(&buf[V2::SECRET..])?,
+            secret: AsymmetricSecretKey::from(&buf[..V2::SECRET_KEY])?,
+            public: AsymmetricPublicKey::from(&buf[V2::SECRET_KEY..])?,
         };
         buf.iter_mut().zeroize();
 
@@ -127,7 +130,7 @@ impl TryFrom<String> for AsymmetricSecretKey<V3> {
     type Error = Error;
 
     fn try_from(value: String) -> Result<Self, Self::Error> {
-        let buf = validate_paserk_string(value.as_str(), "k3", "secret", V3::SECRET)?;
+        let buf = validate_paserk_string(value.as_str(), "k3", "secret", V3::SECRET_KEY)?;
         let ret = Self {
             bytes: buf,
             phantom: PhantomData,
@@ -141,9 +144,9 @@ impl FormatAsPaserk for AsymmetricKeyPair<V4> {
     fn fmt(&self, write: &mut dyn Write) -> core::fmt::Result {
         write.write_str("k4.secret.")?;
 
-        let mut buf = [0u8; V4::SECRET + V4::PUBLIC];
-        buf[..V4::SECRET].copy_from_slice(self.secret.as_bytes());
-        buf[V4::SECRET..].copy_from_slice(self.public.as_bytes());
+        let mut buf = [0u8; V4::SECRET_KEY + V4::PUBLIC_KEY];
+        buf[..V4::SECRET_KEY].copy_from_slice(self.secret.as_bytes());
+        buf[V4::SECRET_KEY..].copy_from_slice(self.public.as_bytes());
         write.write_str(&encode_b64(buf).map_err(|_| core::fmt::Error)?)?;
         buf.iter_mut().zeroize();
 
@@ -155,11 +158,15 @@ impl TryFrom<String> for AsymmetricKeyPair<V4> {
     type Error = Error;
 
     fn try_from(value: String) -> Result<Self, Self::Error> {
-        let mut buf =
-            validate_paserk_string(value.as_str(), "k4", "secret", V4::SECRET + V4::PUBLIC)?;
+        let mut buf = validate_paserk_string(
+            value.as_str(),
+            "k4",
+            "secret",
+            V4::SECRET_KEY + V4::PUBLIC_KEY,
+        )?;
         let ret = Self {
-            secret: AsymmetricSecretKey::from(&buf[..V4::SECRET])?,
-            public: AsymmetricPublicKey::from(&buf[V4::SECRET..])?,
+            secret: AsymmetricSecretKey::from(&buf[..V4::SECRET_KEY])?,
+            public: AsymmetricPublicKey::from(&buf[V4::SECRET_KEY..])?,
         };
         buf.iter_mut().zeroize();
 
@@ -179,7 +186,7 @@ impl TryFrom<String> for AsymmetricPublicKey<V2> {
 
     fn try_from(value: String) -> Result<Self, Self::Error> {
         Ok(Self {
-            bytes: validate_paserk_string(value.as_str(), "k2", "public", V2::PUBLIC)?,
+            bytes: validate_paserk_string(value.as_str(), "k2", "public", V2::PUBLIC_KEY)?,
             phantom: PhantomData,
         })
     }
@@ -197,7 +204,7 @@ impl TryFrom<String> for AsymmetricPublicKey<V3> {
 
     fn try_from(value: String) -> Result<Self, Self::Error> {
         Ok(Self {
-            bytes: validate_paserk_string(value.as_str(), "k3", "public", V3::PUBLIC)?,
+            bytes: validate_paserk_string(value.as_str(), "k3", "public", V3::PUBLIC_KEY)?,
             phantom: PhantomData,
         })
     }
@@ -215,7 +222,7 @@ impl TryFrom<String> for AsymmetricPublicKey<V4> {
 
     fn try_from(value: String) -> Result<Self, Self::Error> {
         Ok(Self {
-            bytes: validate_paserk_string(value.as_str(), "k4", "public", V4::PUBLIC)?,
+            bytes: validate_paserk_string(value.as_str(), "k4", "public", V4::PUBLIC_KEY)?,
             phantom: PhantomData,
         })
     }